fix(tool-catalog): commit Round 1 PR #1032 review fixes left in working tree
The Round 1 review-feedback fixes for the unified tool catalog were made in a
prior session but never committed/pushed, so the remote PR still carried the
unfixed code. A follow-up code review (posted against the stale remote diff)
re-flagged all the same items. This commit lands the already-verified fixes.

catalog.ts (lib/tools/catalog/catalog.ts):
- Finding #1 (latent bug): treat both 'code' and 'retired' rows as code-managed
  when building inactiveCodeKeys. Previously a 'retired' row (a code tool removed
  from the manifest) leaked into the non-code entries list and its key never
  reached inactiveCodeKeys, so a later manifest re-add would spuriously project
  the tool as active even after an admin DB disable.
- Finding #2: documented that invalidate() only detaches the in-flight dbPromise
  reference and does not cancel the underlying query; clarifies the extra-round-
  trip-under-concurrent-invalidation behavior is correctness-safe.
- Finding #4 (security observability): bumped the "tool not in catalog; passing
  through unscoped" log from debug to info so the scope-gate bypass rate is
  observable in production.
- Finding #6 (DRY): version-ordering logic now imported from the shared
  lib/tools/catalog/utils.ts instead of duplicated.

utils.ts (lib/tools/catalog/utils.ts, new):
- Extracted versionRank / compareVersionsDesc shared by manifest.ts and
  catalog.ts (Finding #6).

manifest.ts (lib/tools/catalog/manifest.ts):
- Finding #6: import compareVersionsDesc from ./utils, drop the local copy.

jsonrpc-handler.ts (lib/mcp/jsonrpc-handler.ts):
- Finding #7: add explicit `await` to the now-async handleToolsList call to make
  intent clear.

route.ts (app/api/nexus/chat/route.ts):
- Finding #8: collectAttachmentTexts extracted with its own JSDoc; the
  scanAttachmentPII JSDoc now correctly precedes scanAttachmentPII (no dangling
  comment block).

jsonrpc-catalog.test.ts (tests/unit/lib/mcp/jsonrpc-catalog.test.ts):
- Finding #10: derive expected MCP tool count from TOOL_MANIFEST (MCP_TOOL_COUNT)
  instead of a hardcoded 5, so adding an MCP tool does not break the test.

Findings #3 (dispatch perf), #5 (scope-before-surface ordering) and #9 (extra
deactivation test) were explicitly noted by the reviewer as non-actionable /
correct-as-is / for a future workstream.

Verification: 26/26 catalog unit tests pass (jest), typecheck clean, eslint clean
on all touched files.

Addresses review comments in PR #1032
